The development is working.
From the Jumping System, the ground tiles are located in Level Map >> Missing Prefab
While in the Scenes Ann Goal System, the ground tiles were located in Level Map >> Ground.
Through the merge, the beginning part of the platform (Ground) is in Level Map >> Ground; the rest in Level Map >> Missing Prefab

In every Scene there is a GameCTRl, but inactivation doesn't change the game

Added to the cat GameObjectEnitity, that was the reason it didn't work before

Changed feet to position (X: -0.15; y -2.25 Z:0) and BoxWidth to 0.33.

Deleted PlayerManager class (commented out didn't change the game) and MenuCTRL (not used in changing the scenes)